2009-07-28 12:24 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * include/hbsocket.h   * contrib/hbssl/hbssl.c   * source/rtl/hbsocket.c   * source/rtl/hbinet.c   * examples/uhttpd2/socket.c   * examples/httpsrv/socket.c     * HB_SOCKET_T -> HB_SOCKET
  * contrib/gtwvg/gtwvg.c
  * contrib/gtwvg/wvggui.c
  * source/rtl/gtstd/gtstd.c
  * source/rtl/filesys.c
  * source/rtl/gtpca/gtpca.c
  * source/rtl/gtwvt/gtwvt.c
    ! Added casts for msvc64.
      Some C RTL functions in msvc64 expects file handles as 'int's.
      The rest is some non-critical stuff.
      Now only size_t to int/ULONG warnings remain in hbrtl.
    ; TOFIX: Plus these two:
      ../../../hbstrfmt.c(459) : warning C4244: 'argument' : conversion from '__int64' to 'ULONG', possible loss of data
      ../../../hbtoken.c(280) : warning C4244: '=' : conversion from '__int64' to 'ULONG', possible loss of data

+++ harbour/source/rtl/gtpca/gtpca.c
@@ -644,10 +644,10 @@ static int hb_gt_pca_ReadKey( PHB_GT pGT, int iEventMask )
             ch = s_keyTransTbl[ ch ];
       }
    }
-   else if( !_eof( s_hFilenoStdin ) )
+   else if( !_eof( ( int ) s_hFilenoStdin ) )
    {
       BYTE bChar;
-      if( _read( s_hFilenoStdin, &bChar, 1 ) == 1 )
+      if( _read( ( int ) s_hFilenoStdin, &bChar, 1 ) == 1 )
          ch = s_keyTransTbl[ bChar ];
    }
 #elif defined( HB_IO_WIN )
